third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/adoption.window-expected.txt

This is a testharness.js-based test.
[FAIL] Adoption with explicit global registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Adoption with explicit global registry into a scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with global registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with explicit global registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with global registry into a scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with explicit global registry into a scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
[FAIL] Adoption with explicit global registry (null registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption with explicit global registry into a scoped registry (null registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with global registry (null registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with explicit global registry (null registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with global registry into a scoped registry (null registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with explicit global registry into a scoped registry (null registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption with global registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption with explicit global registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption with global registry into a scoped registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption with explicit global registry into a scoped registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with global registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with explicit global registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with global registry into a scoped registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Adoption including shadow root with explicit global registry into a scoped registry (scoped registry target)
  assert_equals: expected null but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
